pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sana"
version = "0.2.0"
description = "SANA"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "pre-commit==4.5.0",
    "huggingface-hub==0.36.0",
    "accelerate==1.0.1",
    "beautifulsoup4",
    "bs4",
    "came-pytorch",
    "einops",
    "ftfy",
    "wheel",
    "psutil",
    "ninja",
    "diffusers@git+https://github.com/huggingface/diffusers.git",
    "clip@git+https://github.com/openai/CLIP.git",
    "gradio",
    "image-reward",
    "ipdb",
    "mmcv==1.7.2",
    "omegaconf",
    "opencv-python",
    "optimum",
    "patch_conv",
    "peft==0.17.0",
    "protobuf",
    "pytorch-fid",
    "regex",
    "sentencepiece",
    "tensorboard",
    "tensorboardX",
    "timm==0.6.13",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "transformers==4.57.0",
    "triton==3.4.0",
    "wandb",
    "webdataset",
    "xformers==0.0.32.post2",
    "yapf",
    "spaces",
    "matplotlib",
    "termcolor",
    "pyrallis",
    "bitsandbytes",
    "fire",
    "moviepy",
    "imageio[pyav,ffmpeg]",
    "qwen-vl-utils",
    "lmdb",
    "absl-py",
    "ml_collections",
    "hpsv2",
    "open_clip_torch",
]


[project.scripts]
sana-run = "sana.cli.run:main"
sana-upload = "sana.cli.upload2hf:main"

[project.optional-dependencies]

[project.urls]

[tool.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu128"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*", "data*", "output*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*", "data*", "output*"]
